requirements.txt
flask
flask-cors
tensorflow
mediapipe
numpy
opencv-python
requests
librosa
soundfile
moviepy
psycopg2-binary
python-dotenv
h5py
PyJWT
bcrypt
